Backend/requirements.txt
asgiref==3.11.1
asttokens==3.0.1
decorator==5.2.1
Django==5.2.12
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
executing==2.2.1
ipython==9.10.1
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.2.1
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.20.0
PyJWT==2.12.1
python-decouple==3.8
sqlparse==0.5.5
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.15.0
wcwidth==0.6.0
